rafactor: moved classes to main module
todo:
- provideLottoTickets parameter Int > Strategy
- create MockTicketsProvider w/ ManualInputStrategy
- complete implementing test codes
