+++ src/main/kotlin/lotto/domain/LottoTickets.kt
@@ -1,0 +1,3 @@
+package lotto.domain
+
+data class LottoTickets(val lottoTicketList: List<LottoTicket>)
