+++ src/main/kotlin/lotto/provider/ticket/LottoTicketsProvider.kt
@@ -1,0 +1,9 @@
+package lotto.provider.ticket
+
+import lotto.domain.LottoTickets
+
+interface LottoTicketsProvider {
+    fun provideLottoTickets(ticketCount: Int): LottoTickets
+
+    fun provideLottoPrice(): Int
+}
