+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,0 +1,9 @@
+package lotto.view
+
+import lotto.domain.WinningNumber
+
+interface InputView {
+    fun provideBudget(): Int
+
+    fun provideWinningNumbers(): WinningNumber
+}
